Set up Typescript
