+++ .prettierignore
@@ -23,6 +23,7 @@ package-lock.json
 /build/index.html
 /build*
 /build/*
+
 # Editor/IDE
 .idea
 packages/api/etc/
