+++ package.json
@@ -20,6 +20,11 @@
     "npm": ">=3"
   },
   "dependencies": {
+    "@types/jest": "^28.1.1",
+    "@types/node": "^17.0.42",
+    "@types/react": "^17.0.1",
+    "@types/react-dom": "^17.0.1",
+    "@types/styled-components": "^5.1.26",
     "axios": "0.21.4",
     "core-js": "3.2.1",
     "history": "4.10.1",
@@ -55,8 +60,10 @@
     "@babel/plugin-transform-runtime": "7.13.10",
     "@babel/preset-env": "7.13.10",
     "@babel/preset-react": "7.12.13",
+    "@babel/preset-typescript": "^7.18.6",
     "@babel/register": "7.13.8",
     "@babel/runtime": "7.13.10",
+    "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.34.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-loader": "8.2.2",
@@ -97,7 +104,8 @@
     "style-loader": "1.0.0",
     "svg-react-loader": "0.4.6",
     "terser-webpack-plugin": "4.2.3",
-    "typescript": "^4.7.4",
+    "ts-loader": "^8.2.0",
+    "typescript": "^4.6.4",
     "url-loader": "4.1.1",
     "webpack": "4.41.0",
     "webpack-cli": "3.3.9",
