--- src/components/Footer/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Footer = () => <div>Footer</div>;
-
-export default Footer;
