+++ src/components/Footer/index.tsx
@@ -1,0 +1,5 @@
+import React, { FunctionComponent } from 'react';
+
+const Footer: FunctionComponent = () => <footer>Footer</footer>;
+
+export default Footer;
