--- src/components/Header/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Header = () => <div>Header</div>;
-
-export default Header;
