+++ src/components/Header/index.tsx
@@ -1,0 +1,5 @@
+import React, { FunctionComponent } from 'react';
+
+const Header: FunctionComponent = () => <header>Header</header>;
+
+export default Header;
