+++ src/pages/About/index.tsx
@@ -1,0 +1,9 @@
+import React, { FunctionComponent } from 'react';
+
+const About: FunctionComponent = () => (
+  <main>
+    <h1>About Page</h1>
+  </main>
+);
+
+export default About;
