--- src/pages/Home/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Home = () => <h1>Home Page</h1>;
-
-export default Home;
