+++ src/pages/Home/index.tsx
@@ -1,0 +1,9 @@
+import React, { FunctionComponent } from 'react';
+
+const Home: FunctionComponent = () => (
+  <main>
+    <h1>Home Page</h1>
+  </main>
+);
+
+export default Home;
