+++ webpack/paths.js
@@ -3,7 +3,7 @@ const path = require('path');
 module.exports = {
   root: path.resolve(__dirname, '../'),
   outputPath: path.resolve(__dirname, '../', 'build'),
-  entryPath: path.resolve(__dirname, '../', 'src/index.jsx'),
+  entryPath: path.resolve(__dirname, '../', 'src/index.tsx'),
   templatePath: path.resolve(__dirname, '../', 'src/index.html'),
   imagesFolder: 'images',
   fontsFolder: 'fonts',
