+++ webpack/webpack.common.js
@@ -18,6 +18,7 @@ module.exports = {
           sourceMap: true
         }
       },
+      { test: /\.(ts|tsx)$/, loader: 'ts-loader', exclude: [/node_modules/] },
       {
         test: /\.(png|jpg|gif|svg)$/,
         use: [
@@ -67,7 +68,7 @@ module.exports = {
   },
   resolve: {
     modules: ['src', 'node_modules'],
-    extensions: ['*', '.js', '.jsx', '.css', '.scss']
+    extensions: ['*', '.js', '.jsx', '.ts', '.tsx', '.css', '.scss']
   },
   plugins: [
     new webpack.ProgressPlugin(),
